refactor(ViewBuilder): Add primitives to create a view container, provider and call to a functional directive.
This is CL is part of multipart refactor to make createMethod private to enable
optimizations.

PiperOrigin-RevId: 174091753
