+++ _goldens/test_files/di/generated_injectors.template_release.golden
